feat: implementar utils para mejor reutilizacion del codigo
